+++ examples/plot_healpix.py
@@ -11,7 +11,7 @@ data. Also feel free to turn ``with_timing`` back on.
 """
 
 import cogsworth
-p = cogsworth.pop.Population(1000, processes=1, BSE_settings={"binfrac": 1.0},
+p = cogsworth.pop.Population(1000, processes=1,
                              use_default_BSE_settings=True)
 p.create_population(with_timing=False)